Revert "Herramientas"
This reverts commit be3ce1474e989eb42e75719b7dd3bb12e40eebaa.

+++ BattleRoyale/src/ProyectoFinalJava/Herramientas.java
@@ -3,10 +3,10 @@ package ProyectoFinalJava;
 public class Herramientas {
 
 	private String nombre;
-	private Double bonus;
+	private Integer bonus;
 	private TipoArma tipo;
 	
-	public Herramientas(String nombre, Double bonus, TipoArma tipo) {
+	public Herramientas(String nombre, Integer bonus, TipoArma tipo) {
 		this.nombre = nombre;
 		this.bonus = bonus;
 		this.tipo = tipo;
@@ -20,11 +20,11 @@ public class Herramientas {
 		this.nombre = nombre;
 	}
 
-	public Double getBonus() {
+	public Integer getBonus() {
 		return bonus;
 	}
 
-	public void setBonus(Double bonus) {
+	public void setBonus(Integer bonus) {
 		this.bonus = bonus;
 	}
